requirements.txt
branca==0.3.1
certifi==2019.11.28
chardet==3.0.4
Django==2.2.6
django-geojson==2.12.0
django-tables2==2.1.1
djangorestframework==3.10.3
folium==0.10.0
idna==2.8
Jinja2==2.10.3
jsonfield==2.0.2
MarkupSafe==1.1.1
mysqlclient==1.4.4
numpy==1.18.1
pandas==0.25.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.10
sqlparse==0.3.0
urllib3==1.25.8
xlrd==1.2.0
